Add Chrome tracing for callChromeHiddenMethod
Extensions tend to spend a lot of their time in JavaScript invoked from here.  This change is to help see that time is being spend in javascript in chrome://tracing.

This is related to crbug.com/112892 which tracks the fact that the extension system calling directly into V8 here also misses a bunch of other instrumentation we normally get when entering JavaScript.

BUG=114017

Review URL: http://codereview.chromium.org/9389011

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@122859 0039d316-1c4b-4281-b951-d872f2087c98