Allow for sync integration tests to restart the sync service
One sync scenario we don't test yet is where a browser is restarted
after having signed in to sync, and sync should just start up on its
own.

This patch adds functionality to the integration test framework that
allows for the sync service to be restared with the same user data
directory, exactly as if all browser windows for a given profile were
closed and reopened.

It also adds a simple test case that restarts the sync service and makes
sure sync is up and running after the restart.

BUG=112644
TEST=SingleClientBookmarksSyncTest.RestartSyncService

Review URL: https://chromiumcodereview.appspot.com/9317091

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@123533 0039d316-1c4b-4281-b951-d872f2087c98