Revert 124071 - Gestures are now possible using touch events with any ids
Reverting because this CL caused any tap in a tab contents (on device) to
trigger a browser crash.  See issue 116314 for details.

--

Each gesture point stores a "point_id," which is used for state transitions in the GestureRecognizer.
The point_ids are maintained such that the set of point_ids is always contiguous, from 0 to the number of current touches.
A lower point_id indicates that a touch occurred first.

BUG=113144
TEST=GestureRecognizerTest.*

TBR=tdresser@chromium.org
Review URL: https://chromiumcodereview.appspot.com/9566020

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@124451 0039d316-1c4b-4281-b951-d872f2087c98