Suppress key-repeat on Linux host
This ensures that key-repeat is controlled by the client, not the host (so that
network latency will not trigger key-repeat on the host if the release event is
delayed).  To test this, disable key-repeat on the client machine (on Linux,
run "xset r off").  Then verify that this works over a connection to a Linux
host (key presses from the client are not auto-repeated on the host).

This implementation should work for both IT2Me and Me2Me.  In the IT2Me case,
the local user of the host machine should experience key-repeats normally.
The host's global preference for auto-repeat is not permanently changed; the
auto-repeat is only suppressed whilst keys are pressed from the client.


BUG=115972
TEST=Manual


Review URL: http://codereview.chromium.org/9614011

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@125288 0039d316-1c4b-4281-b951-d872f2087c98