bluetooth: implement application-level connections
On a successful connection, or on a reconnection to a previously paired
or connected device, introspect the D-Bus object and make the appropriate
application-level connection calls to connect the device to the kernel.

Mark the device trusted after successful connection so it can be reconnected
after reboot or device power cycle.

At the same time merge Connect and PairAndConnect, there's no reason for
the UI code to need to care about the difference; especially if we in the
future support connection without pairing for certain device classes.

BUG=chromium-os:21322
TEST=tested on device

Change-Id: Ibdeeab77127893708f8ad367a2f6cb348ea25b60

Review URL: https://chromiumcodereview.appspot.com/9704047

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@127049 0039d316-1c4b-4281-b951-d872f2087c98