Make SyncManagerTests use in-memory sqlite
Add a flag to the SyncManager to allow it to initialize an in-memory
directory.  This is to be used only in tests.

This is an ugly change, but there's not much alternative available right
now.  We really ought to figure out a better way to test the SyncManager
that doesn't require cluttering the main class with flag variables.
That task is left to another day.

Using an in-memory database results in significant speedups.  The time
taken to run sync_unit_tests on my desktop went from 13s to 6s.  It
seems to cut 10-20s off the runtime of unit_test
--gtest_filter='*Sync*', too.

BUG=116328, 117836
TEST=


Review URL: http://codereview.chromium.org/9662019

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@127255 0039d316-1c4b-4281-b951-d872f2087c98