bluetooth: add support for Passkey pairing
Bluetooth 2.1 onwards defines a pairing mechanism where a device
without capability for Input or Display, and that declines Passkey-less
pairing for security reasons, can require us to enter a six-digit
numeric Passkey to perform pairing.

This will, of course, never happen.

But in case it does, implement support for it. This requires us knowing
whether the device is expecting a PIN Code or Passkey, and in the latter
case converting the String we receive from JavaScript into a numeric to
return.

BUG=chromium-os:21320
TEST=verified with a fake device

Change-Id: I0fe420f8fab7073ab8c158eb372ec227f7fa53f5

R=satorux1@chromium.org

Review URL: https://chromiumcodereview.appspot.com/9704074

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@127506 0039d316-1c4b-4281-b951-d872f2087c98