Fix schema_bundle_generator's capitalization of function names.
Python's String.capitalize() function ensures that only the first character in
the returned string is a captial. For multi-word extension functions this is
incorrect (e.g. "createContext" becomes "Createcontext"). This patch changes
the behaviour of schema_bundle_generator to only attempt to captialize the
first character of the function name.

BUG=119295
TEST=


Review URL: http://codereview.chromium.org/9818002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@128024 0039d316-1c4b-4281-b951-d872f2087c98