Implement lazy JS module initialisation.
The existing lazy JS object code only allows single objects to be evaluated
lazily, but a common pattern in existing modules is to define multiple objects
and functions in the same closure and attach them as various fields on an
existing object, which that interface doesn't provide for. The approach
implemented in this change leverages the module system by allowing clients to
set 'lazy fields' on already existing objects, which when accessed evaluate to
a given field in a given module.


BUG=104100
TEST=unit tests provided


Review URL: http://codereview.chromium.org/9812025

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@128131 0039d316-1c4b-4281-b951-d872f2087c98