Ash: Special header transparency for single open windows
Added a special transparency setting for when a single large window is open.
Made window transparency values based on constants, rather than baked into assets.
Added chrome://transparency so designers can fiddle with the values without waiting for an engineer to make a change and the builds to cycle.  We'll take this out when we have final values.

BUG=120275,120454
TEST=manual
R=sky@chromium.org

Review URL: https://chromiumcodereview.appspot.com/9864010

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@129266 0039d316-1c4b-4281-b951-d872f2087c98