cros: Fix crash when clicking on discarded tab
In out of memory conditions Chrome OS R19 will "discard" tabs that have not been recently used.  The page is reloaded when the user selects the tab in the tab strip.  We're seeing occasional crashes in the field in the tab reload code due to null site instance pointers in the navigation controller.  This might have been caused by crrev.com/120007 which added the use of site instance.  I have no in-house repro for this crash but I have a user who can trigger it reliably -- we'll see if this patch helps her.

BUG=119068
TEST=crash reports should go down

Review URL: https://chromiumcodereview.appspot.com/9802047

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@129940 0039d316-1c4b-4281-b951-d872f2087c98