Remove installer strings from chrome.dll.
They are only used by the installer (the header installer_util_strings.h
is only included in files in chrome/installer), so they don't
need to be in chrome.dll.

It looks like this was added in r1044, but I don't think it's
needed anymore.

BUG=119947

Review URL: https://chromiumcodereview.appspot.com/9809036

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@129947 0039d316-1c4b-4281-b951-d872f2087c98