Revert 128179 - Make sure the plugin scriptable object is released before NPP_Destroy.
We're temporarily reverting this change to gather more data on its impact on plugin crash rates.

When the we tear down a plugin instance the plugin process first invokes NPP_Destroy, and then tears down the IPC channel to the renderer, to give NPP_Destroy a chance to do last-minute scripting.  When the IPC channel for the last instance is torn down we also clean up the IPC channels and stubs for any plugin-side NPObjects that remain.

We suspect that some plugins implement the scriptable object as part of the plugin instance, rather than independently ref-counted, so that our releasing the object after NPP_Destroy actually triggers the plugin process to crash.

This CL tears down the stub for the plugin's scriptable object before we call NPP_Destroy.

As per crbug.com/119414, we will remove this code if it doesn't significantly impact crashes.

BUG=101968

Review URL: http://codereview.chromium.org/9817023

TBR=cpu@chromium.org
Review URL: https://chromiumcodereview.appspot.com/9959078

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@130199 0039d316-1c4b-4281-b951-d872f2087c98