Revert 130190 - Move InMemoryURLIndex Caching Operations to FILE Thread
Reading and writing of the InMemoryURLIndex (IMUI) cache is now peformed on the FILE thread and, when required, rebuilding of the index is performed on the history thread. Reading and rebuilding of the index from the cache file or history database are now performed as atomic operations rather than by restoring/rebuilding in-place, i.e. a new private data object is created and populated then swapped in for the live private data object.

BUG=83659
TEST=Unit tests updated/enhanced.

Previously reviewed: https://chromiumcodereview.appspot.com/9030031 (This change identical to 9030031.)
Review URL: https://chromiumcodereview.appspot.com/9703124

TBR=mrossetti@chromium.org
Review URL: https://chromiumcodereview.appspot.com/9969071

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@130206 0039d316-1c4b-4281-b951-d872f2087c98