GTK Panels fix for height glitch after opening.
Auto-resizable panels expanded slightly in height when mouse is
moved over a UI element in the panel after first creation and
when re-activated from overflow. This is because the frame
size is reduced if the panel is sized smaller than the frame,
such as the initial minimum size of 100x20 for auto-resizable
panels, or the overflow panel size.

Resolved by storing the frame size from when it was first allocated.

BUG=None
TEST=Open an auto-resizable panel and mouse over its content. No height change.

Review URL: https://chromiumcodereview.appspot.com/9784015

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@131208 0039d316-1c4b-4281-b951-d872f2087c98