Switch PepperView to use asynchronous completion for Flush().
PepperView was requesting "optional" synchronous/asynchronous completion from Flush(), so as to be able to handle errors, which was not possible with the CompletionCallbackClosureAdapter, since the closure would not receive the result.

This CL replaces CompletionCallbackClosureAdapter with a PpCompletionCallback() function that accepts a base::Callback<void(int)> and returns a pp::CompletionCallback suitable for passing directly to Flush().  PepperView is also updated to use the new helper, and to expect Flush() to always complete asynchronously.

TEST=Chromoting client renders correctly.


Review URL: http://codereview.chromium.org/10010033

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@131304 0039d316-1c4b-4281-b951-d872f2087c98